+++ src/test/java/scootertests/DropdownTextTest.java
@@ -1,99 +1,65 @@
 package scootertests;
 
 import com.example.scooters.HomePageScooter;
-
-import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 
-public class DropdownTextTest {
-    HomePageScooter objHomePage;
-    WebDriver driver;
 
-    @Before
-    public void before() {
-        //драйвер для браузера Chrome
-        //создание экземпляра драйвера
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--no-sandbox", "--headless", "--disable-dev-shm-usage");
-        driver = new ChromeDriver(options);
-        // переход на страницу тестового приложения
-        driver.get("https://qa-scooter.praktikum-services.ru/");
-        // Найди раздел «Вопросы о важном» и сделать скролл до неё
-        WebElement element = driver.findElement(By.cssSelector(".Home_FAQ__3uVm4"));
-        ((JavascriptExecutor) driver).executeScript("arguments[0].scrollIntoView();", element);
-        // создай объект класса главной страницы приложения
-        objHomePage = new HomePageScooter(driver);
-    }
+public class DropdownTextTest extends TestBase {
 
-    @Test
-    public void testOpen() {
-        HomePageScooter homePageScooter = new HomePageScooter(driver);
-        homePageScooter.open();
-    }
     @Test
             public void acceptCookie(){
         HomePageScooter homePageScooter = new HomePageScooter(driver);
         homePageScooter.clickAcceptCookie();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
     @Test
-    public void checkAccordion0() {
-        objHomePage.clickPanel(0);
-        Assert.assertTrue(objHomePage.isPanelVisible(0));
+    public void checkHowMuch() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion1() {
-        objHomePage.clickPanel(1);
-        Assert.assertTrue(objHomePage.isPanelVisible(1));
+    public void checkMoreScooters() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion2() {
-        objHomePage.clickPanel(2);
-        Assert.assertTrue(objHomePage.isPanelVisible(2));
+    public void checkTimeRent() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion3() {
-        objHomePage.clickPanel(3);
-        Assert.assertTrue(objHomePage.isPanelVisible(3));
+    public void checkRentScootersNow() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion4() {
-        objHomePage.clickPanel(4);
-        Assert.assertTrue(objHomePage.isPanelVisible(4));
+    public void checkExtendScooter() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion5() {
-        objHomePage.clickPanel(5);
-        Assert.assertTrue(objHomePage.isPanelVisible(5));
+    public void checkGiveChargeWithScooters() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion6() {
-        objHomePage.clickPanel(6);
-        Assert.assertTrue(objHomePage.isPanelVisible(6));
+    public void checkCancelOrder() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
     @Test
-    public void checkAccordion7() {
-        objHomePage.clickPanel(7);
-        Assert.assertTrue(objHomePage.isPanelVisible(7));
+    public void checkILiveBehindMkhad() {
+        objHomePage.clickPanel();
+        Assert.assertTrue(objHomePage.isPanelVisible());
     }
 
-    @After
-    public void after() {
-        // Закрыть браузер
-        driver.quit();
-    }
+
 }
